1) MessageProcessorLib: for handling the received message and preparing the corresponding response, 2) HwSettingsLib: CANData: is extended to hold the info for both received and Tx message, 3) UtilityLib: HighPrecisionTimer is implemented for handing the receiving/sending of CAN messages with microsecond intervals, 4) Start Transmission functionality is implemented - using two instances of HighPrecisionTimer to receive (from the BUS) and send (corresponding response from the messageConfig.json) the CAN/CANFD messages

+++ VectorXLWrapper/VectorUtil.cs
@@ -46,9 +46,9 @@ namespace HardwareDriverLayer.Wrapper
         /// </summary>
         /// <param name="dlc"></param>
         /// <returns></returns>
-        internal static uint GET_CANFD_PAYLOAD(XLDefine.XL_CANFD_DLC dlc)
+        internal static byte GET_CANFD_PAYLOAD(XLDefine.XL_CANFD_DLC dlc)
         {
-            uint payload = 8;
+            byte payload = 8;
             switch (dlc)
             {
                 case XLDefine.XL_CANFD_DLC.DLC_CAN_CANFD_0_BYTES: payload = 0; break; //DLC = 0
